refactor: notification 타입 정리
