+++ src/notifications/interfaces/notification.service.interface.ts
@@ -1,4 +1,4 @@
-import { Notification } from '#notifications/notification.types.js';
+import { Notification } from '#notifications/types/notification.types.js';
 
 export interface INotificationService {
   getNotifications: (page: number, pageSize: number) => Promise<{ totalCount: number; list: Notification[] }>;
